requirements.txt
# Django Core
Django==5.0.1
django-environ==0.11.2

# MongoDB Integration
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4

# API & Serialization
djangorestframework==3.14.0

# CORS Headers (for API)
django-cors-headers==4.3.1

# Development Tools
python-decouple==3.8

# Data Processing (for movie datasets)
pandas==2.1.4
numpy==1.26.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0
